// src/redux/common/types.ts
export enum ActionType {
  INIT_APP = 'COMMON/INIT_APP',
  INIT_APP_SUCCESS = 'COMMON/INIT_APP_SUCCESS',
  ERROR = 'COMMON/ERROR',
  INFO = 'COMMON/INFO',
  SUCCESS = 'COMMON/SUCCESS',
  DISMISS_MESSAGE = 'COMMON/DISMISS_MESSAGE',
  PREVIEW_IMAGES = 'COMMON/PREVIEW_IMAGES',
  PREVIEW_END = 'COMMON/PREVIEW_END',
  SET_TOKEN = 'COMMON/SET_TOKEN',
  LOAD_AllSITE = 'COMMON/LOAD_ALL_SITE',
  LOAD_AllSITE_SUCCESS = 'COMMON/LOAD_ALL_SITE_SUCCESS',
  LOAD_CITY = 'COMMON/LOAD_CITY',
  LOAD_CITY_SUCCESS = 'COMMON/LOAD_CITY_SUCCESS',
}